refactor: make tts config more understandable and easier
The TTS config now only has two instead of three booleans to manage caching.
It now always saves the cached audio files in StreamingAssets (there was no alternative before)
and the users can decide whether they want to also cache during runtime when the application
was already built or not.
This is useful when the TTS provider is not always available (e.g. no Internet connection).
Android cannot write into StreamingAssets, so this should be deactivated when building for Android.